Responding to warnings raised by gcc-3.3
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1814 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/RenderOther/FontHandler.cpp
@@ -334,7 +334,7 @@ void FontSpecifier::OGL_Reset(bool IsStarting)
  		{
  			MoveTo(HPos,VPos);
  			DrawChar(Which);
- 			HPos += widths_p[Which++];
+ 			HPos += widths_p[static_cast<int>(Which++)];
  		}
  	}
